requirements.txt
PyYAML==5.3.1
python-box==5.2.0
flask==1.1.2
flask-restful==0.3.8
